reworked silabs main.cpp to follow template. Moved functionality to matterconfig.cpp
