+++ examples/platform/silabs/MatterConfig.h
@@ -26,6 +26,7 @@ class SilabsMatterConfig
 {
 public:
     static CHIP_ERROR InitMatter(const char * appName);
+    static void app_init();
 
 private:
     static CHIP_ERROR InitOpenThread(void);
