+++ src/platform/silabs/platformAbstraction/GsdkSpam.cpp
@@ -44,7 +44,6 @@ extern "C" {
 
 #include "sl_component_catalog.h"
 #include "sl_mbedtls.h"
-#include "sl_system_init.h"
 #if SILABS_LOG_OUT_UART || ENABLE_CHIP_SHELL || CHIP_DEVICE_CONFIG_THREAD_ENABLE_CLI
 #include "uart.h"
 #endif
@@ -68,8 +67,6 @@ SilabsPlatform::SilabsButtonCb SilabsPlatform::mButtonCallback = nullptr;
 
 CHIP_ERROR SilabsPlatform::Init(void)
 {
-    sl_system_init();
-
     mRebootCause = RMU_ResetCauseGet();
     // Clear register so it does accumualate the causes of each reset
     RMU_ResetCauseClear();
